% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippie-package.R, R/hippie.R
\name{hippie-invoke}
\alias{hippie-invoke}
\alias{hippie}
\alias{hippie_up}
\alias{hippie_down}
\title{Hippie code completion}
\usage{
hippie_up()

hippie_down()
}
\value{
A match token is inserted into the relevant RStudio source document
context.
}
\description{
Invoke hippie completion, either upwards or downwards.
}
