% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippie-package.R, R/hippie.R
\name{hippie-invoke}
\alias{hippie-invoke}
\alias{hippie}
\alias{hippie_up}
\alias{hippie_down}
\title{Hippie Code Completion}
\usage{
hippie_up()

hippie_down()
}
\value{
Nothing. A match token is inserted into the relevant RStudio source
document context.
}
\description{
Invoke Hippie Completion, either upwards or downwards. This function is meant
to be called as a shortcut.
}
\examples{
\dontrun{
# Not intended to be called directly. Rather, bind to a keyboard shortcut.
hippie_up()
hippie_down()
}
}
